Allfiles/Labs/18/requirements.txt
azure-functions>=1.20.0,<2.0.0
psycopg[binary]>=3.2.1,<4.0.0
